move pos device detail component
