+++ src/components/paytacapos/PosDeviceDetailDialog.vue
